Fountains require tears of armok and provide water.
Fountains are no longer decorative places only useful for swimming.
Using a tear of armok, a pipe section, and a mechanism, a fountain
can be created which allows for buckets and bottles to be filled.

This means that Orcs and Kobolds now have a use for the tear
of armok, as they don't normally have access to the alchemist.
This also provides an alternative way for dwarves to generate
water from a tear.