creature_birds.txt and creature_subterranean.txt files now use APPLY_CREATURE_VARIATION.
There were differences between how masterwork and other tilesets
constructured birds. This patch brings all of these into line with
each other, and is important in getting BillyJack's less leather
patch to apply.

No other changes are made. In particular it was noted that
ostriches are trainable in some tilesets, but not others.