Automatic harmonization of graphics raws with master files.
Modified all tilesets to match the maste files where diferences were
detected in *non-display* elements.

This is an automatic harmonization on changes which were detected
to be safe, and constitutes approximately 30% of the detected
significant, non-display differences between tilesets and the master
files.

The remaining ~70% of mismatches require either human intervention,
or (more likely) smarter code to harmonize them.