Masterwork 4f release
Since this is based upon the `MWDF_4f_release_candidate_r4` tagged commit,
I've imported the history of that up until this point on the upstream
branch.

This commit brings things to be inline exactly as they are with Meph's
released file (which doesn't have the maint tools or source code).