removing filenames
the loading time is down to ~6.5 seconds, which is a significant
improvement. some of the civ options are still broken as they require
tags in the raws to make each option unique.

if the graphic pack raws are combined as discussed (multiple lines for
tile changes) then things can only improve as it'll be ~400 less files
that have to be read! in the interim, it may be better to add a splash
screen while the initial load happens.